+++ CHANGELOG.md
@@ -58,6 +58,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Update integration tests to verify multiple instances of SqlDatabaseMail
     in the same configuration ([issue #1871](https://github.com/dsccommunity/SqlServerDsc/issues/1871)).
   - Changed comparison logic to use `Compare-DscParameterState` ([issue #1871](https://github.com/dsccommunity/SqlServerDsc/issues/1871)).
+  - Return the correct profile name and mail server name from the current
+    state ([issue #1871](https://github.com/dsccommunity/SqlServerDsc/issues/1871)).
 - `Invoke-SqlDscQuery`
   - Now shows the correct instance name when called using a server object
     ([issue #1918](https://github.com/dsccommunity/SqlServerDsc/issues/1918)).